Fix rollover from normal key to glukey
This change makes it so that rollover from a non-glukey to a glukey doesn't set the glukey
state to `clear` when the normal key is released. This is done by preventing the setting
of `release_trigger_` unless there are currently any glukeys with a `temp` bit set.

In order to do this, we need to keep a count of the number of bits set in `temp_state_[]`,
by using the `temp_state_count_` variable.

Fixes #9

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>