Prevent modifiers & layer change keys from becoming release trigger
Modifiers and layer keys should not prevent glukeys from being used as `sticky` with other
keys. For example: tap glukey `ctrl`, hold `shift`, tap `S`. The result should be
`ctrl`+`shift`+`S`, but without this change the `control` key would get released as soon
as the `S` keypress event was recieved, before it gets sent to the OS in the report.

Fixes #11

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>