Fix rollover after sticky layer-shift
If a `sticky` layer-shift glukey is in effect, it should only apply to the next key
toggled on, which means that the layer shift must be cancelled immediately after the
trigger key is pressed. Prior to this change, it was only being cancelled on release of
the trigger key, or when rollover occurred to some other key. In the rollover case, the
value of that next key should not come from the shifted layer, and the simplest way to
achieve that is by releasing the layer-shift key as soon as the trigger key is pressed.

Fixes #14

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>