Added a "meta-glukey" that can turn other keys into glukeys
The meta-glukey state is tracked using the address of the active meta-glukey, and is
itself a glukey, but needs special treatment because it needs to be released (when
`sticky`) by _any_ keypress, regardless of type. I haven't yet tested a keymap with two
meta-glukeys, so there might still be bugs in that corner case.

Closes #12

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>