Fix layer-shift glukeys releasing prematurely when held
If a layer-shift glukey got locked, `layer_shift_addr_` would persist. If the same key was
then pressed and held, it would get released prematurely as soon as the next key was
pressed. This change clears `layer_shift_addr_` when that address enters the `locked`
state to prevent this from happening.

Fixes #20

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>